requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
# The != are from recent oslo.log
Babel!=2.3.0,!=2.3.1,!=2.3.2,!=2.3.3,>=1.3 # BSD
SQLAlchemy<1.1.0,>=1.0.10 # MIT
oslo.db>=4.1.0 # Apache-2.0
oslo.versionedobjects>=1.5.0 # Apache-2.0
oslo.config>=3.7.0 # Apache-2.0
oslo.policy>=1.15.0 # Apache-2.0
oslo.log>1.14.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.middleware>=3.0.0
six>=1.9.0 # MIT
WSME>=0.8 # MIT
pecan>=1.0.0 # BSD
requests!=2.9.0,>=2.8.1 # Apache-2.0
rfc3986>=0.3.1 # Apache-2.0
PyYAML>=3.1.0 # MIT
pytz>=2013.6 # MIT
click>=6.6
python-etcd>=0.4.3
keystonemiddleware!=4.5.0,>=4.2.0
keystoneauth1>=2.14.0 # Apache-2.0